Create Remove-recursive-file.ps1
"Remove Malicious Archives Script for Compromised Machines"

This script is designed to assist in the removal of malicious archives from compromised machines. It will take a user-specified argument and recursively delete all files and directories associated with. This can be useful for quickly and effectively removal of files, directories of a compromised system.